Add optional baseDir parameter
The optional <baseDir> configuration parameter
makes it possible to restrict
where docbook-linktester looks for files.

This is an alternative to using <excludes>
when source files
have raw Maven properties to filter
or share names with files to exclude,
such as src/site/xdoc/index.xml.

It might actually make more sense
to figure out how to do this with <excludes>.

Tested with forgerock-doc-maven-plugin-test